requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
httpx
faiss-cpu
sentence-transformers
langchain
pypdf